Move document references from Search to SDDocumentType
